chore(frontend): move helpers to separate folders tm-388
