+++ apps/frontend/src/pages/chats/libs/helpers/get-formatted-chat-date/get-formatted-chat-date.helper.ts
