+++ apps/frontend/src/pages/chats/libs/helpers/helpers.ts
@@ -1,2 +1,2 @@
-export { getFormattedChatDate } from "./get-formatted-chat-date.helper.js";
-export { prepareMessageItems } from "./prepare-message-items.helper.js";
+export { getFormattedChatDate } from "./get-formatted-chat-date/get-formatted-chat-date.helper.js";
+export { prepareMessageItems } from "./prepare-message-items/prepare-message-items.helper.js";
